+++ packages/espressocash_app/lib/features/activities/module.dart
@@ -5,9 +5,7 @@ import 'package:provider/provider.dart';
 
 import '../../core/accounts/bl/account.dart';
 import '../../core/accounts/module.dart';
-import '../../core/balances/context_ext.dart';
 import '../../di.dart';
-import '../swap/swap_verifier.dart';
 import 'src/updater/bloc.dart';
 
 class ActivitiesModule extends SingleChildStatelessWidget {
@@ -21,12 +19,6 @@ class ActivitiesModule extends SingleChildStatelessWidget {
             create: (context) => sl<TxUpdaterBloc>(
               param1: context.read<MyAccount>().wallet,
             ),
-          ),
-          Provider<SwapVerifier>(
-            lazy: false,
-            create: (context) => sl<SwapVerifier>()
-              ..init(onBalanceAffected: () => context.notifyBalanceAffected()),
-            dispose: (_, value) => value.dispose(),
           ),
         ],
         child: LogoutListener(