+++ packages/espressocash_app/lib/features/incoming_split_key_payments/src/widgets/incoming_split_key_payment_screen.dart
@@ -1,8 +1,7 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../../core/balances/context_ext.dart';
+import '../../../../core/transactions/tx_sender.dart';
 import '../../../../di.dart';
 import '../../../../l10n/l10n.dart';
 import '../../../../ui/transfer_status/transfer_error.dart';
@@ -10,8 +9,8 @@ import '../../../../ui/transfer_status/transfer_progress.dart';
 import '../../../../ui/transfer_status/transfer_success.dart';
 import '../../widgets/invalid_escrow_error_widget.dart';
 import '../bl/incoming_split_key_payment.dart';
-import '../bl/iskp_bloc.dart';
 import '../bl/iskp_repository.dart';
+import 'extensions.dart';
 
 class IncomingSplitKeyPaymentScreen extends StatefulWidget {
   const IncomingSplitKeyPaymentScreen({
@@ -42,32 +41,31 @@ class _IncomingSplitKeyPaymentScreenState
         builder: (context, state) {
           final payment = state.data;
 
-          return BlocConsumer<ISKPBloc, ISKPState>(
-            listener: (context, state) => payment?.status.mapOrNull(
-              txSent: (_) => context.notifyBalanceAffected(),
+          if (payment == null) {
+            return TransferProgress(
+              onBack: () => context.router.pop(),
+            );
+          }
+
+          return payment.status.maybeMap(
+            success: (_) => TransferSuccess(
+              onBack: () => context.router.pop(),
+              onOkPressed: () => context.router.pop(),
+              statusContent: context.l10n.moneyReceived,
             ),
-            builder: (context, state) {
-              if (payment == null || state.contains(payment.id)) {
-                return TransferProgress(
-                  onBack: () => context.router.pop(),
-                );
+            txFailure: (it) {
+              if (it.reason == TxFailureReason.escrowFailure) {
+                return const InvalidEscrowErrorWidget();
               }
 
-              return payment.status.maybeMap(
-                success: (_) => TransferSuccess(
-                  onBack: () => context.router.pop(),
-                  onOkPressed: () => context.router.pop(),
-                  statusContent: context.l10n.moneyReceived,
-                ),
-                txEscrowFailure: (_) => const InvalidEscrowErrorWidget(),
-                orElse: () => TransferError(
-                  onBack: () => context.router.pop(),
-                  onRetry: () => context
-                      .read<ISKPBloc>()
-                      .add(ISKPEvent.process(payment.id)),
-                ),
+              return TransferError(
+                onBack: () => context.router.pop(),
+                onRetry: () => context.retryISKP(payment),
               );
             },
+            orElse: () => TransferProgress(
+              onBack: () => context.router.pop(),
+            ),
           );
         },
       );