+++ packages/espressocash_backend/espressocash_api/lib/src/dto/create_payment.dart
@@ -69,6 +69,7 @@ class CreateDirectPaymentResponseDto with _$CreateDirectPaymentResponseDto {
   const factory CreateDirectPaymentResponseDto({
     required int fee,
     required String transaction,
+    required BigInt slot,
   }) = _CreateDirectPaymentResponseDto;
 
   factory CreateDirectPaymentResponseDto.fromJson(Map<String, dynamic> json) =>